requirements.txt
paho-mqtt==1.3.1
psycopg2-binary==2.7.4
python-dateutil==2.6.1
SQLAlchemy==1.2.4
Flask==0.12.2
